+++ index.js
@@ -2,8 +2,7 @@ require('dotenv').config()
 const Discord = require('discord.js')
 const { prefix } = require('./config.json')
 const { displayError } = require('./helpers/error')
-const { getRandomActivity } = require('./helpers/activityChanger')
-const { init } = require('./helpers/init')
+const { getRandomActivity, init } = require('./helpers')
 const { initCommands } = require('./data/Commands')
 
 const client = new Discord.Client()
